:sparkles:(api): Database Updated models: move them all in one file & edit messages
